Add WarpStation Cap (deltaGiga+baseWarp) feature.
Stop making warpstations if we are past the deltagiga + base
warpstations (and no giga upgrade is available). Will also remove the
green highlight around the icon. This will save you metal to use on
weapons,armor, etc.
NOTE: (the cap will only work on incremental buys, it will not come into
effect when the game uses a gigastation and immediately bulk-buys as
many warpstations as it can afford. I think this is preferrable.)