Dynamic Siphonology - only when needed based on (Enemyhealth / baseDamage)
Created a new setting in the advanced options. "Dynamic Siphonology".
It will switch to the proper Map-level as soon as the current map is
completed.
So you can choose original behavior of always using the lowest level
map, or the modified behavior, which increases the map level based on
your damage.
The old behavior of "no siphonology at all when using DisableFarming" is
no longer applied, under any circumstance.