Add Corruption handling for 2 of the corruption types (Strong and Tough).
TODO: deal with repetition and coagulate all the "enemy" determinations into a single function that doesnt happen in autostance specifically so that automaps can rely on it, and autoLevelEquipment as well. It gets tricky because autostance responds on an individual cell by cell case basis, but the other two seem to be over-all Zone by zone basises.