Auto Upgrade Heirlooms
Made a lot of Code to make the script work on starting from a fresh savegame (Delete Save) 0 Helium , Level 1, Zone 1 etc.....
Update some equations for getEnemyMaxHealth and Attack (nearly identical, except for difficulty before level 60 was adjusted in some patch a couple months ago and never accounted for). Also, now you can call them with a "corrupt = true" parameter if you want to.
Adjust storage buying so that the script cannot buy a storage building before it is unlocked at level 1 and 70%.