He/Hr Dont Portal Before.
Do NOT allow Helium per Hour AutoPortal setting to portal BEFORE this level is reached. It is an additional check that prevents drops in helium/hr from triggering autoportal. Set to 0 or -1 to completely disable this check.

unrelated: double check to make sure the nursery is actually still locked before bothering to try to run a map to unlock it.