BugFix - Update gather.js so Watch and Watch2 dont' get stuck because of Trapping/noTrimps.
Long story:
The problem with watch is that AT hires everything, yet at the end of the zone the game also hires them automatically (thus watch), so you can end up without trimps to breed and be stuck.

Theoretically trapping could solve this, but it was only building traps and not actually checking the traps to help get out of the cycle.